refactor: Provider에서 hook 분리
