+++ src/providers/UserProvider/index.tsx
@@ -2,7 +2,6 @@ import {
   PropsWithChildren,
   createContext,
   useCallback,
-  useContext,
   useEffect,
   useMemo,
   useState,
@@ -22,9 +21,10 @@ type UserInfoActionsContext = {
   signOut: () => void;
 } | null;
 
-const UserInfoContext = createContext<UserInfoContext>(null);
+export const UserInfoContext = createContext<UserInfoContext>(null);
 
-const UserInfoActionsContext = createContext<UserInfoActionsContext>(null);
+export const UserInfoActionsContext =
+  createContext<UserInfoActionsContext>(null);
 
 const UserProvider = ({ children }: PropsWithChildren) => {
   const { pathname } = useLocation();
@@ -84,28 +84,3 @@ const UserProvider = ({ children }: PropsWithChildren) => {
 };
 
 export default UserProvider;
-
-export const useUserInfo = () => {
-  const value = useContext(UserInfoContext);
-
-  if (value === null) {
-    return {
-      isLoading: true,
-    };
-  }
-
-  return {
-    data: value,
-    isLoading: false,
-  };
-};
-
-export const useUserInfoAction = () => {
-  const value = useContext(UserInfoActionsContext);
-
-  if (value === null) {
-    throw new Error('UserInfoActionsContext 에러');
-  }
-
-  return value;
-};
