+++ src/types/user.ts
@@ -1,9 +1,9 @@
 export type UserInfo = {
   email: string;
   nickname: string;
-  school: {
-    code: string;
-    areaCode: string;
-    name: string;
+  school?: {
+    code?: string;
+    areaCode?: string;
+    name?: string;
   };
 };
